refactor: Enable strict Rust lints
nplan-id: enable-strict-rust-lints
nplan-decision: [dec:nsh:bytes-not-text]
nplan-decision: [dec:nsh:differential-is-the-oracle]
nplan-decision: [dec:nsh:errors-are-values]
nplan-decision: [dec:nsh:fork-child-is-a-terminus]
nplan-decision: [dec:nsh:host-owns-signals]
nplan-decision: [dec:nsh:host-owns-streams]
nplan-decision: [dec:nsh:host-owns-the-process]
nplan-decision: [dec:nsh:minimal-unsafe]
nplan-decision: [dec:nsh:no-ambient-state]
nplan-decision: [dec:nsh:no-format-interpreters]
nplan-decision: [dec:nsh:owned-data]
nplan-decision: [dec:nsh:per-shell-locale]
nplan-decision: [dec:nsh:printf-is-parsed-not-interpreted]
nplan-decision: [dec:nsh:public-surface]
nplan-decision: [dec:nsh:shell-as-library]
nplan-decision: [dec:nsh:we-own-the-defects]
nplan-element: [arch:nsh:shell-core]
nplan-rule: [spec:dash:sem:expand.addfname-common-fn]
nplan-rule: [spec:dash:sem:expand.addfnamealt-fn]
nplan-rule: [spec:dash:sem:expand.addglob-fn]
nplan-rule: [spec:dash:sem:expand.argstr-fn]
nplan-rule: [spec:dash:sem:expand.arith-fn]
nplan-rule: [spec:dash:sem:expand.chtodest-fn]
nplan-rule: [spec:dash:sem:expand.cvtnum-fn]
nplan-rule: [spec:dash:sem:expand.esclen-fn]
nplan-rule: [spec:dash:sem:expand.evalvar-fn]
nplan-rule: [spec:dash:sem:expand.expandmeta-fn]
nplan-rule: [spec:dash:sem:expand.expandmeta-glob-fn]
nplan-rule: [spec:dash:sem:expand.expari-fn]
nplan-rule: [spec:dash:sem:expand.expbackq-fn]
nplan-rule: [spec:dash:sem:expand.expcmd-fn]
nplan-rule: [spec:dash:sem:expand.expmeta-fn]
nplan-rule: [spec:dash:sem:expand.expmeta-rmescapes-fn]
nplan-rule: [spec:dash:sem:expand.expsort-fn]
nplan-rule: [spec:dash:sem:expand.exptilde-fn]
nplan-rule: [spec:dash:sem:expand.mbtodest-fn]
nplan-rule: [spec:dash:sem:expand.memtodest-fn]
nplan-rule: [spec:dash:sem:expand.mesclen-fn]
nplan-rule: [spec:dash:sem:expand.msort-fn]
nplan-rule: [spec:dash:sem:expand.opendir-interruptible-fn]
nplan-rule: [spec:dash:sem:expand.preglob-fn]
nplan-rule: [spec:dash:sem:expand.removerecordregions-fn]
nplan-rule: [spec:dash:sem:expand.rmescapes-fn]
nplan-rule: [spec:dash:sem:expand.scanleft-fn]
nplan-rule: [spec:dash:sem:expand.scanright-fn]
nplan-rule: [spec:dash:sem:expand.strtodest-fn]
nplan-rule: [spec:dash:sem:expand.subevalvar-fn]
nplan-rule: [spec:dash:sem:expand.varunset-fn]
nplan-rule: [spec:dash:sem:expand.varvalue-fn]
nplan-rule: [spec:dash:sem:histedit.setterm-fn]
nplan-rule: [spec:dash:sem:myhistedit.setterm-fn]
nplan-rule: [spec:dash:sem:shell.max-int-length-fn]
nplan-rule: [spec:nsh:def:idiom.shell-options]
nplan-rule: [spec:nsh:req:compat.smoosh.error-contracts]
nplan-rule: [spec:nsh:req:idiom.lexer-tokens]
nplan-rule: [spec:nsh:req:idiom.parser-control-flow]
nplan-rule: [spec:nsh:req:idiom.strict-lints]
nplan-rule: [spec:nsh:sem:idiom.specified-defects]
nplan-rule: [spec:posix:def:exit.expansion-error]
nplan-rule: [spec:posix:def:expand.tilde-prefix-in-assignment]
nplan-rule: [spec:posix:def:expand.tilde-prefix]
nplan-rule: [spec:posix:def:param.positional-definition]
nplan-rule: [spec:posix:def:param.special-parameters]
nplan-rule: [spec:posix:def:pattern.filename-expansion-qualification]
nplan-rule: [spec:posix:req:expand.arith-token-expansion]
nplan-rule: [spec:posix:req:expand.cmdsub-no-reexpansion]
nplan-rule: [spec:posix:req:expand.cmdsub-semantics]
nplan-rule: [spec:posix:req:expand.param-assign-default]
nplan-rule: [spec:posix:req:expand.param-colon-effect]
nplan-rule: [spec:posix:req:expand.param-error-if-unset]
nplan-rule: [spec:posix:req:expand.param-hash-requires-word]
nplan-rule: [spec:posix:req:expand.param-remove-largest-prefix]
nplan-rule: [spec:posix:req:expand.param-remove-largest-suffix]
nplan-rule: [spec:posix:req:expand.param-remove-smallest-prefix]
nplan-rule: [spec:posix:req:expand.param-remove-smallest-suffix]
nplan-rule: [spec:posix:req:expand.param-simple]
nplan-rule: [spec:posix:req:expand.param-string-length]
nplan-rule: [spec:posix:req:expand.param-substring-common]
nplan-rule: [spec:posix:req:expand.param-use-alternative]
nplan-rule: [spec:posix:req:expand.param-use-default]
nplan-rule: [spec:posix:req:expand.param-word-expansion]
nplan-rule: [spec:posix:req:expand.pathname]
nplan-rule: [spec:posix:req:expand.quote-removal]
nplan-rule: [spec:posix:req:expand.tilde-home]
nplan-rule: [spec:posix:req:expand.tilde-login-name]
nplan-rule: [spec:posix:req:expand.tilde-replacement-pathname]
nplan-rule: [spec:posix:req:expand.tilde-result-quoted]
nplan-rule: [spec:posix:req:param.positional-decimal-digits]
nplan-rule: [spec:posix:req:param.special-asterisk]
nplan-rule: [spec:posix:req:param.special-at-double-quotes]
nplan-rule: [spec:posix:req:param.special-at-no-positional]
nplan-rule: [spec:posix:req:param.special-at]
nplan-rule: [spec:posix:req:param.special-bang]
nplan-rule: [spec:posix:req:param.special-dollar]
nplan-rule: [spec:posix:req:param.special-hash]
nplan-rule: [spec:posix:req:param.special-hyphen]
nplan-rule: [spec:posix:req:param.special-question]
nplan-rule: [spec:posix:req:param.special-zero]
nplan-rule: [spec:posix:req:pattern.directory-permissions]
nplan-rule: [spec:posix:req:pattern.filename-expansion-trigger]
nplan-rule: [spec:posix:req:pattern.leading-period-in-bracket-unspecified]
nplan-rule: [spec:posix:req:pattern.leading-period]
nplan-rule: [spec:posix:req:pattern.no-match-unchanged]
nplan-rule: [spec:posix:req:pattern.no-special-chars-unchanged]
nplan-rule: [spec:posix:req:pattern.permission-errors-not-fatal]
nplan-rule: [spec:posix:req:pattern.replacement-sorted]
nplan-rule: [spec:posix:req:pattern.slash-explicit-match]
nplan-rule: [spec:posix:req:pattern.unmatched-open-bracket-unspecified]
nplan-rule: [spec:posix:sem:expand.quote-removal-quoting-remembered]
nplan-rule: [spec:posix:sem:expand.tilde-no-further-expansion]
nplan-rule: [spec:posix:sem:param.positional-zero-not-positional]
nplan-rule: [spec:posix:sem:param.special-question-assignment]
nplan-rule: [spec:posix:syn:expand.param-braces-optional]
nplan-rule: [spec:posix:syn:expand.param-format]
nplan-rule: [spec:posix:syn:expand.param-unbraced-resolution]
nplan-rule: [spec:posix:syn:param.positional-multi-digit-braces]
nplan-rule: [spec:posix:syn:pattern.slash-terminates-bracket]
nplan-rule: [spec:nsh:req:idiom.strict-lints/test]

+++ crates/nsh/src/parser/bash.rs
@@ -239,7 +239,7 @@ pub(super) fn process_substitutions(
             return Ok(());
         }
 
-        lexer.output.push(LEGACY_COMMAND_SUBSTITUTION as u8);
+        lexer.output.push(LEGACY_COMMAND_SUBSTITUTION);
         let parked = mem::take(&mut lexer.output);
         let slot = lexer.command_substitutions.len();
         lexer.command_substitutions.push(None);
@@ -675,12 +675,12 @@ fn matching_bracket(bytes: &[u8], open: usize) -> Option<usize> {
     let mut quoted = false;
     while index < bytes.len() {
         match bytes[index] {
-            byte if byte == LEGACY_ESCAPE as u8 => index += 2,
-            byte if byte == LEGACY_QUOTE as u8 => {
+            byte if byte == LEGACY_ESCAPE => index += 2,
+            byte if byte == LEGACY_QUOTE => {
                 quoted = !quoted;
                 index += 1;
             }
-            byte if byte == LEGACY_MULTIBYTE as u8 => {
+            byte if byte == LEGACY_MULTIBYTE => {
                 index = multibyte_end(bytes, index);
             }
             b'[' if !quoted => {
@@ -732,12 +732,12 @@ fn backquote_count(bytes: &[u8]) -> usize {
     let mut count = 0usize;
     let mut index = 0usize;
     while index < bytes.len() {
-        if bytes[index] == LEGACY_ESCAPE as u8 {
+        if bytes[index] == LEGACY_ESCAPE {
             index += 2;
-        } else if bytes[index] == LEGACY_MULTIBYTE as u8 {
+        } else if bytes[index] == LEGACY_MULTIBYTE {
             index = multibyte_end(bytes, index);
         } else {
-            count += usize::from(bytes[index] == LEGACY_COMMAND_SUBSTITUTION as u8);
+            count += usize::from(bytes[index] == LEGACY_COMMAND_SUBSTITUTION);
             index += 1;
         }
     }
@@ -745,7 +745,7 @@ fn backquote_count(bytes: &[u8]) -> usize {
 }
 
 fn multibyte_end(bytes: &[u8], start: usize) -> usize {
-    let length_at = start + 1 + usize::from(bytes.get(start + 1) == Some(&(LEGACY_ESCAPE as u8)));
+    let length_at = start + 1 + usize::from(bytes.get(start + 1) == Some(&(LEGACY_ESCAPE)));
     let length = bytes.get(length_at).copied().unwrap_or(0) as usize;
     length_at.saturating_add(length).saturating_add(3)
 }
